[기능수정] 헤더 컴포넌트 로그인 로그아웃 버튼 분리
